Move ProgressBar into it's own file.
